requirements.txt
flet>=0.21.0
onnxruntime>=1.16.0
librosa>=0.10.0
sounddevice>=0.4.6
numpy>=1.24.0
matplotlib>=3.7.0
pandas>=2.0.0
Pillow>=10.0.0
scipy>=1.11.0
psutil>=5.9.0
